move setActiveEnvironment to wrapper
